\name{biodiv_df}

\alias{biodiv_df}

\title{Example biodiversity data.frame}

\description{
Example biodiversity data.frame
}

\usage{
data(biodiv_df)
}

\details{
Example of input \code{biodiv_df} used for functions.
}

\examples{
\dontrun{
data(biodiv_df)
head(biodiv_df)
}
}

\value{
data.frame object with information about features.
}